Refactored sealed hierarchy and fixed SQlite tests
